Verififed!
Yes, I think we did it. Now the next step is line bundling. But first, we need to modularize this piece of code an beautify it.
Also, there is no other job to be updated, so I guess this can be shown to mam as the final output of the first process.
Please confirm.